add: UserReview 변경 및 get, insert 구현

+++ back/tripDuo/src/main/java/com/example/tripDuo/dto/UserProfileInfoDto.java
@@ -34,7 +34,7 @@ public class UserProfileInfoDto {
     
     private String socialLinks; // > json 처리
     
-    private Float ratings; // 지표 설정 
+    private Long ratings; // 지표 설정 
 
     private String lastLogin; // 몇분전 접속
